+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Version History
 ===============
 
+v5.16.0
+--------
+
+* Update GIS component to adapt with newest CSC version `<https://github.com/lsst-ts/LOVE-frontend/pull/441>`_
+* Observatory Summary Component Implementation `<https://github.com/lsst-ts/LOVE-frontend/pull/440>`_
+
 v5.15.1
 --------